runners/rexfw/pyproject.toml
[tool.poetry]
name = "resaas-runner-rexfw"
version = "0.1.0"
description = "RESAAS runner for Replica Exchange sampling with rexfw"
authors = ["simeoncarstens <simeon.carstens@tweag.io>"]
packages = [ {include = "resaas" } ]

[tool.poetry.dependencies]
python = "^3.8"
numpy = "*"
pyyaml = "*"
resaas-common = { path = "../../common", develop = true }
rexfw = { git = "https://github.com/tweag/rexfw", branch = "start-script" }
mpi4py = "^3.0.3"

[tool.poetry.dev-dependencies]
pytest = "^6.2.1"
black = "^20.8b1"
flake8 = "^3.8.4"
pylint = "^2.6.0"
mypy = "^0.790"
pytest-dotenv = "^0.5.2"
pytest-cov = "^2.10.1"
isort = "^5.7.0"
jedi = "*"
rope = "*"

[tool.black]
line-length = 99
target-version = ["py38"]
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
minversion = "6.0"
env_override_existing_values = 1
env_files = [".env"]

[tool.coverage.run]
omit = ["*test*"]
